backend/requirements.txt
fastapi>=0.115.5
uvicorn>=0.30.0
pydantic-ai>=0.0.14
pydantic>=2.8.0
redis>=5.0.8
asyncpg>=0.29.0
sqlalchemy>=2.0.35
alembic>=1.13.2
logfire>=0.50.0
httpx>=0.27.0
python-dotenv>=1.0.1
anthropic>=0.34.0
duckduckgo-search>=6.2.6
mcp>=1.0.0
python-multipart>=0.0.9
websockets>=13.0
aiosqlite>=0.20.0
